# openapi.yaml
openapi: 3.1.0
info:
  title: Predictify API
  version: 0.0.1
  description: Backend API for Predictify — a Stellar/Soroban prediction-markets dApp
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            requestId:
              type: string
          required:
            - code
      required:
        - error
    ValidationErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            details:
              nullable: true
          required:
            - code
      required:
        - error
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        stellarAddress:
          type: string
        score:
          type: number
      required:
        - rank
        - stellarAddress
        - score
...
    UserListRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stellarAddress:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - stellarAddress
        - createdAt
    CurrentUserProfile:
      type: object
      properties:
        stellarAddress:
          type: string
        createdAt:
          type: string
          format: date-time
        totals:
          type: object
          properties:
            prediction_count:
              type: integer
            claim_count:
              type: integer
          required:
            - prediction_count
            - claim_count
      required:
        - stellarAddress
        - createdAt
        - totals
    Prediction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        marketId:
          type: string
        status:
          type: string
          enum: &ref_0
            - pending
            - confirmed
            - won
            - lost
            - claimed
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - marketId
        - status
        - createdAt
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stellarAddress:
          type: string
        joinedAt:
          type: string
          format: date-time
        predictions:
          type: array
          items:
            $ref: '#/components/schemas/Prediction'
        totals:
          type: object
          properties:
            prediction_count:
              type: integer
            claim_count:
              type: integer
          required:
            - prediction_count
            - claim_count
      required:
        - id
        - stellarAddress
        - joinedAt
        - predictions
        - totals
...
paths:
...
  /api/leaderboard/user/{stellarAddress}:
    get:
      operationId: getLeaderboardUser
      tags:
        - Leaderboard
      summary: Get leaderboard entry for a specific user
      parameters:
        - schema:
            type: string
          required: true
          name: stellarAddress
          in: path
      responses:
        '200':
          description: Entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/LeaderboardEntry'
                required:
                  - data
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/users:
    get:
      operationId: listUsers
      tags:
        - Users
      summary: List all users (cursor-paginated)
      description: >-
        Returns a cursor-paginated list of registered users sorted newest-first (DESC createdAt, DESC id).  Pass the
        opaque `nextCursor` value from one response as the `cursor` query parameter of the next request to advance
        through pages.  A null `nextCursor` indicates the last page. Supports strong ETag / conditional GET: send the
        ETag back as If-None-Match on subsequent requests; if the page is unchanged the server responds 304 Not Modified
        (no body).
      parameters:
        - schema:
            type: string
            description: Opaque cursor token from the previous page's nextCursor field.
            example: djF8MjZ8MjAyNi0wMS0wMVQwMDowMDowMC4wMDBafGFiY2Qtd...
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            description: Page size (1–100, default 20).
          required: false
          name: limit
          in: query
        - schema:
            type: string
            description: ETag from a previous 200 response. Triggers 304 when the page is unchanged.
          required: false
          name: If-None-Match
          in: header
      responses:
        '200':
          description: Paginated list of users
          headers:
            ETag:
              description: Strong ETag (SHA-256) of the response body.
              schema:
                type: string
            Cache-Control:
              description: Always no-cache so clients revalidate before reuse.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserListRow'
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor for the next page, or null on the last page.
                required:
                  - data
                  - nextCursor
        '304':
          description: Not Modified — page unchanged since the ETag in If-None-Match.
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
  /api/users/me:
    get:
      operationId: getCurrentUser
      tags:
        - Users
      summary: Get the authenticated user’s profile
      description: >-
        Returns the authenticated user's profile. Supports strong ETag / conditional GET: send the ETag back as
        If-None-Match on subsequent requests; if the profile is unchanged the server responds 304 Not Modified (no
        body).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            description: ETag from a previous 200 response. Triggers 304 when content is unchanged.
          required: false
          name: If-None-Match
          in: header
      responses:
        '200':
          description: Current user profile
          headers:
            ETag:
              description: Strong ETag (SHA-256) of the response body.
              schema:
                type: string
            Cache-Control:
              description: Always no-cache so clients revalidate before reuse.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CurrentUserProfile'
                required:
                  - data
              examples:
                currentUser:
                  value:
                    data:
                      stellarAddress: GABC1234567890DEFGHIJKLMNOPQRSTUVWX
                      createdAt: '2026-06-27T12:00:00.000Z'
                      totals:
                        prediction_count: 2
                        claim_count: 0
        '304':
          description: Not Modified — profile unchanged since the ETag in If-None-Match.
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                unauthorized:
                  value:
                    error:
                      code: UNAUTHORIZED
                      requestId: req_xyz789
  /api/users/{address}/predictions:
    get:
      operationId: getUserPredictions
      tags:
        - Users
      summary: List predictions for a Stellar address
      description: >-
        Returns a cursor-paginated list of predictions. Supports strong ETag / conditional GET: send the ETag back as
        If-None-Match; if the page is unchanged the server responds 304 Not Modified (no body).
      parameters:
        - schema:
            type: string
          required: true
          name: address
          in: path
        - schema:
            type: string
            enum: *ref_0
          required: false
          name: status
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
            description: ETag from a previous 200 response. Triggers 304 when the page is unchanged.
          required: false
          name: If-None-Match
          in: header
      responses:
        '200':
          description: Paginated predictions
          headers:
            ETag:
              description: Strong ETag (SHA-256) of the response body.
              schema:
                type: string
            Cache-Control:
              description: Always no-cache so clients revalidate before reuse.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Prediction'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
              examples:
                samplePage:
                  value:
                    data:
                      - id: 11111111-1111-1111-1111-111111111111
                        marketId: market-abc-123
                        status: confirmed
                        createdAt: '2026-06-27T12:00:00.000Z'
                    nextCursor: djF8MjR8...
        '304':
          description: Not Modified — predictions page unchanged since the ETag in If-None-Match.
        '400':
          description: Invalid address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                invalidAddress:
                  value:
                    error:
                      code: invalid_address
                      requestId: req_abc123
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                notFound:
                  value:
                    error:
                      code: not_found
                      requestId: req_abc123
  /api/users/{stellarAddress}/profile:
    get:
      operationId: getUserProfile
      tags:
        - Users
      summary: Get a user’s public profile
      description: >-
        Returns a public user profile. Supports strong ETag / conditional GET: send the ETag back as If-None-Match; if
        the profile is unchanged the server responds 304 Not Modified (no body).
      parameters:
        - schema:
            type: string
          required: true
          name: stellarAddress
          in: path
        - schema:
            type: string
            description: ETag from a previous 200 response. Triggers 304 when the profile is unchanged.
          required: false
          name: If-None-Match
          in: header
      responses:
        '200':
          description: User profile
          headers:
            ETag:
              description: Strong ETag (SHA-256) of the response body.
              schema:
                type: string
            Cache-Control:
              description: Always no-cache so clients revalidate before reuse.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserProfile'
                required:
                  - data
              examples:
                publicProfile:
                  value:
                    data:
                      id: 22222222-2222-2222-2222-222222222222
                      stellarAddress: GXYZ1234567890ABCDEFGHIJKLMNOPQRSTUV
                      joinedAt: '2025-01-01T12:00:00.000Z'
                      predictions:
                        - id: 33333333-3333-3333-3333-333333333333
                          marketId: market-def-456
                          status: won
                          createdAt: '2026-06-27T12:00:00.000Z'
                      totals:
                        prediction_count: 1
                        claim_count: 1
        '400':
          description: Invalid Stellar address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                invalidAddress:
                  value:
                    error:
                      code: validation_error
                      requestId: req_xyz789
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                notFound:
                  value:
                    error:
                      code: not_found
                      requestId: req_xyz789
...
  /api/admin/users/{address}:
    get:
      operationId: getAdminUser
      tags:
        - Admin
      summary: Get aggregated user data for admin support
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: address
          in: path
      responses:
        '200':
          description: Admin user view
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AdminUserView'
                required:
                  - data
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...